VOS/requirements.txt
customtkinter
psutil
py-cpuinfo
requests
pyinstaller
pycaw
comtypes
flask
flask-cors
reportlab
GPUtil
pystray
Pillow
